Module gtest::constants

Module containing constants of Gear protocol.

Constants

BLOCK_DURATION_IN_MSECS

Duration of one block in msecs.

DISPATCH_HOLD_COST

Cost of storing delayed message per block.

EPOCH_DURATION_IN_BLOCKS

Duration of one epoch.

EXISTENTIAL_DEPOSIT

Minimal amount of value able to be sent. Defines accounts existence requirement.

GAS_ALLOWANCE

Gas allowance for executing user dispatch and set of generated by programs dispatches from execution of the user dispatch.

GAS_MULTIPLIER

Gas multiplier used to calculate equivalence of gas in token value.

INITIAL_RANDOM_SEED

Initial random seed for testing environment.

LOAD_ALLOCATIONS_PER_INTERVAL

Cost of loading allocations per interval.

MAILBOX_THRESHOLD

Minimal amount of gas required to be inserted into Mailbox.

MAX_RESERVATIONS

Maximal amount of reservations program may have.

MODULE_CODE_SECTION_INSTANTIATION_BYTE_COST

Cost of wasm module code section instantiation before execution per byte of code.

MODULE_DATA_SECTION_INSTANTIATION_BYTE_COST

Cost of wasm module data section instantiation before execution per byte of code.

MODULE_ELEMENT_SECTION_INSTANTIATION_BYTE_COST

Cost of wasm module element section instantiation before execution per byte of code.

MODULE_GLOBAL_SECTION_INSTANTIATION_BYTE_COST

Cost of wasm module global section instantiation before execution per byte of code.

MODULE_INSTRUMENTATION_BYTE_COST

Cost of instrumenting wasm code on upload per byte of code.

MODULE_INSTRUMENTATION_COST

Cost of instrumenting wasm code on upload.

MODULE_TABLE_SECTION_INSTANTIATION_BYTE_COST

Cost of wasm module table section instantiation before execution per byte of code.

MODULE_TYPE_SECTION_INSTANTIATION_BYTE_COST

Cost of wasm module type section instantiation before execution per byte of code.

READ_COST

Cost of read access into storage.

READ_PER_BYTE_COST

Per-byte extra cost of read access into storage.

RESERVATION_COST

Cost of storing reservation per block.

RESERVE_FOR

Extra amount of blocks must be reserved for storing in storage.

UNITS

Value per token.

VALUE_PER_GAS

Value per gas.

WAITLIST_COST

Cost of storing waitlisted message per block.

WRITE_COST

Cost of write access into storage.

WRITE_PER_BYTE_COST

Per-byte extra cost of write access into storage.

Type Aliases

Block

Numeric type representing blocks in Gear protocol.

Gas

Numeric type representing gas in Gear protocol.

Value

Numeric type representing value in Gear protocol.