gtest

Struct System

pub struct System(/* private fields */);

The testing environment which simulates the chain state and its transactions but somehow the real on-chain execution environment could be different.

use gtest::System;

// Create a new testing environment.
let system = System::new();

// Init logger with "gwasm" target set to `debug` level.
system.init_logger();

Implementations

impl System

pub fn new() -> Self

Create a new testing environment.

Panics

Only one instance in the current thread of the System is possible to create. Instantiation of the other one leads to runtime panic.

pub fn init_logger(&self)

Init logger with “gwasm” target set to debug level.

pub fn init_verbose_logger(&self)

Init logger with “gwasm” and “gtest” targets set to debug level.

pub fn init_logger_with_default_filter<'a>( &self, default_filter: impl Into<Cow<'a, str>>, )

Init logger with default_filter as default filter.

pub fn queue_len(&self) -> usize

Returns amount of dispatches in the queue.

pub fn run_next_block(&self) -> BlockRunResult

Run next block.

Block execution model is the following:

increase the block number, update the timestamp

process tasks from the task pool

process messages in the queue.

The system is always initialized with a 0 block number. Current block number in the system is the number of the already executed block, therefore block execution starts with a block info update (block number increase, timestamp update). For example, if current block number is 2, it means that messages and tasks on 2 were executed, so the method goes to block number 3 and executes tasks and messages for the updated block number.

Task processing basically tries to execute the scheduled to the specific block tasks:

delayed sending

waking message

removing from the mailbox

removing reservations

removing stalled wait message.

Messages processing executes messages until either queue becomes empty or block gas allowance is fully consumed.

pub fn run_next_block_with_allowance(&self, allowance: Gas) -> BlockRunResult

Runs blocks same as Self::run_next_block, but with limited allowance.

pub fn run_to_block(&self, bn: u32) -> Vec<BlockRunResult>

Runs blocks same as Self::run_next_block, but executes blocks to block number bn including it.

pub fn run_scheduled_tasks(&self, amount: u32) -> Vec<BlockRunResult>

Runs amount of blocks only with processing task pool, without processing the message queue.

pub fn block_height(&self) -> u32

Return the current block height of the testing environment.

pub fn block_timestamp(&self) -> u64

Return the current block timestamp of the testing environment.

pub fn get_program<ID: Into<ProgramIdWrapper>>( &self, id: ID, ) -> Option<Program<'_>>

Returns a Program by id.

pub fn last_program(&self) -> Option<Program<'_>>

Returns last added program.

pub fn programs(&self) -> Vec<Program<'_>>

Returns a list of programs.

pub fn is_active_program<ID: Into<ProgramIdWrapper>>(&self, id: ID) -> bool

Detect if a program is active with given id.

An active program means that the program could be called, instead, if returns false it means that the program has exited or terminated that it can’t be called anymore.

pub fn inheritor_of<ID: Into<ProgramIdWrapper>>( &self, id: ID, ) -> Option<ActorId>

Returns Some(ActorId) if a program is exited with inheritor.

Returns None otherwise.

pub fn submit_local_code_file<P: AsRef<Path>>(&self, code_path: P) -> CodeId

Saves code to the storage and returns its code hash

Same as [‘submit_code_file’], but the path is provided as relative to the current directory.

pub fn submit_code_file<P: AsRef<Path>>(&self, code_path: P) -> CodeId

Saves code from file to the storage and returns its code hash

See also System::submit_code

pub fn submit_code(&self, binary: impl Into<Vec<u8>>) -> CodeId

Saves code to the storage and returns its code hash

This method is mainly used for providing a proper program from program creation logic. In order to successfully create a new program with gstd::prog::create_program_bytes_with_gas function, developer should provide to the function “child’s” code hash. Code for that code hash must be in storage at the time of the function call. So this method stores the code in storage.

Also method saves instrumented version of the code.

pub fn submitted_code(&self, code_id: CodeId) -> Option<Vec<u8>>

Returns previously submitted original code by its code hash.

pub fn get_mailbox<ID: Into<ProgramIdWrapper>>( &self, id: ID, ) -> ActorMailbox<'_>

Extract mailbox of user with given id.

The mailbox contains messages from the program that are waiting for user action.

pub fn mint_to<ID: Into<ProgramIdWrapper>>(&self, id: ID, value: Value)

Mint balance to user with given id and value.

pub fn transfer( &self, from: impl Into<ProgramIdWrapper>, to: impl Into<ProgramIdWrapper>, value: Value, keep_alive: bool, )

Transfer balance from user with given from id to user with given to id.

pub fn balance_of<ID: Into<ProgramIdWrapper>>(&self, id: ID) -> Value

Returns balance of user with given id.

pub fn calculate_reply_for_handle( &self, origin: impl Into<ProgramIdWrapper>, destination: impl Into<ProgramIdWrapper>, payload: impl Into<Vec<u8>>, gas_limit: u64, value: Value, ) -> Result<ReplyInfo, String>

Calculate reply that would be received when sending message to initialized program with any of Program::send* methods.