Module gtest::constants

Module containing constants of Gear protocol.

Constants

BLOCK_DURATION_IN_MSECS

Duration of one block in msecs.

DISPATCH_HOLD_COST

Cost of storing delayed message per block.

EPOCH_DURATION_IN_BLOCKS

Duration of one epoch.

EXISTENTIAL_DEPOSIT

Minimal amount of value able to be sent. Defines accounts existence requirement.

INITIAL_RANDOM_SEED

Initial random seed for testing environment.

MAILBOX_THRESHOLD

Minimal amount of gas required to be inserted into Mailbox.

MAX_RESERVATIONS

Maximal amount of reservations program may have.

MODULE_INSTANTIATION_BYTE_COST

Cost of wasm module instantiation before execution per byte of code.

MODULE_INSTRUMENTATION_BYTE_COST

Cost of instrumenting wasm code on upload per byte of code.

MODULE_INSTRUMENTATION_COST

Cost of instrumenting wasm code on upload.

READ_COST

Cost of read access into storage.

READ_PER_BYTE_COST

Per-byte extra cost of read access into storage.

RESERVATION_COST

Cost of storing reservation per block.

RESERVE_FOR

Extra amount of blocks must be reserved for storing in storage.

UNITS

Value per token.

VALUE_PER_GAS

Value per gas.

WAITLIST_COST

Cost of storing waitlisted message per block.

WRITE_COST

Cost of write access into storage.

WRITE_PER_BYTE_COST

Per-byte extra cost of write access into storage.

Type Aliases

Block

Numeric type representing blocks in Gear protocol.

Gas

Numeric type representing gas in Gear protocol.

Value

Numeric type representing value in Gear protocol.