gtest
Module constants

Module containing constants of Gear protocol.

Constants

BLOCK_DURATION_IN_MSECS

Duration of one block in msecs.

DEFAULT_USERS_INITIAL_BALANCE

Default initial balance for users.

DEFAULT_USER_ALICE

Default user id for Alice.

DEFAULT_USER_BOB

Default user id for Bob.

DEFAULT_USER_CHARLIE

Default user id for Charlie.

DEFAULT_USER_EVE

Default user id for Eve.

EPOCH_DURATION_IN_BLOCKS

Duration of one epoch.

EXISTENTIAL_DEPOSIT

Minimal amount of value able to be sent. Defines accounts existence requirement.

GAS_ALLOWANCE

Gas allowance for executing user dispatch and set of generated by programs dispatches from execution of the user dispatch.

GAS_MULTIPLIER

Gas multiplier used to calculate equivalence of gas in token value.

INITIAL_RANDOM_SEED

Initial random seed for testing environment.

MAX_RESERVATIONS

Maximal amount of reservations program may have.

MAX_USER_GAS_LIMIT

Max allowed user gas to apply for external message.

RESERVE_FOR

Extra amount of blocks must be reserved for storing in storage.

UNITS

Value per token.

VALUE_PER_GAS

Value per gas.

Functions

default_users_list

Default list of users.

Type Aliases

Block

Numeric type representing blocks in Gear protocol.

Gas

Numeric type representing gas in Gear protocol.

Value

Numeric type representing value in Gear protocol.